style: rename some folder
