+++ src/Parts/Imu.cpp
@@ -1,4 +1,4 @@
-#include "Driver/Imu.hpp"
+#include "Parts/Imu.hpp"
 
 // BNO055のレジスタアドレスなどを定義
 #define BNO055_ADDRESS (0x28 << 1)
